Get rid of hypothesis
