+++ tests/test_symmetry.py
@@ -1,23 +1,22 @@
-from hypothesis import given
-from hypothesis.strategies import frozensets, integers, lists, one_of, sets, tuples
+from typing import List, Set, Tuple, Optional, FrozenSet
 
-from serious import JsonModel
+from serious import JsonModel, DictModel
 from serious.test_utils import assert_symmetric
-from tests.entities import (DataclassWithFrozenSet, DataclassWithList, DataclassWithOptional,
-                            DataclassWithSet, DataclassWithTuple)
-from tests.hypothesis2 import examples
-from tests.hypothesis2.strategies import optionals
+from tests.entities import dataclass_of
+from tests.utils import with_
 
-dcconss_strategies_conss = [
-    (DataclassWithList, lists, list, [1]),
-    (DataclassWithSet, sets, set, [1]),
-    (DataclassWithTuple, tuples, tuple, [1]),
-    (DataclassWithFrozenSet, frozensets, frozenset, [1]),
-    (DataclassWithOptional, optionals, lambda x: x, 1)
+models = [JsonModel, DictModel]
+objects = [
+    dataclass_of(List[int])([1, 2, 3]),
+    dataclass_of(Set[int])({1, 2, 3}),
+    dataclass_of(Tuple[int, Ellipsis])(tuple([1, 2, 3])),
+    dataclass_of(Tuple[int, str, int])(tuple([1, 'two', 3])),
+    dataclass_of(FrozenSet[int])(frozenset([1, 2, 3])),
+    dataclass_of(Optional[int])(None),
+    dataclass_of(Optional[int])(3),
 ]
 
 
-@given(one_of(*[strategy_fn(integers()).map(dccons) for dccons, strategy_fn, *_ in dcconss_strategies_conss]))
-@examples(*[dccons(cons(input)) for dccons, _, cons, input in dcconss_strategies_conss])
-def test_generic_encode_and_decode_are_inverses(dc):
-    assert_symmetric(JsonModel(type(dc)), dc)
+@with_(models, objects)
+def test_generic_encode_and_decode_are_inverses(new_model, dc):
+    assert_symmetric(new_model(type(dc)), dc)
